moves signee state tag component to its own file
